pyproject.toml
[tool.poetry]
name = "laederhub"
version = "0.1.0"
description = "Self-serve analytics backend MVP for connecting SMB SaaS silos"
authors = ["Nick Laeder <nick@example.com>"]
readme = "README.md"
packages = [{ include = "services" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
python-dotenv = "^1.0.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
firebase-admin = "^6.5.0"
google-cloud-firestore = "^2.14.0"
google-cloud-bigquery = "^3.17.0"
google-cloud-storage = "^2.16.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core>=1.8.2"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
